# client: avoid copying Batch Request slice in DB.Run and Txn.Run