@petermattis 2017-02-18
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/13660#-:-KdEC8OWQoxQlFU4GeOi:bnfp4nl)*
<!-- Sent from Reviewable.io -->